Jwt userId checks updated

+++ src/main/java/com/cs334/project3/api/Controller.java
@@ -61,13 +61,26 @@ public class Controller {
     }
 
     @GetMapping(value="/groups", params= "userId")
-    public ResponseEntity<List<GroupDTO>> getGroupsForUser(@RequestParam Long userId){
-        try {
-            List<GroupDTO> gDTOList = groupMemberService.getGroupsWhereUserIsMember(userId);
-            return new ResponseEntity<>(gDTOList, HttpStatus.OK);
-        } catch (Exception e){
-            e.printStackTrace();
-            throw new InternalServerErrorException("Exception raised trying to retrieve group");
+    public ResponseEntity<List<GroupDTO>> getGroupsForUser(@RequestParam Long userId, HttpServletRequest request){
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if(uid.equals(userId)) {
+            try {
+                List<GroupDTO> gDTOList = groupMemberService.getGroupsWhereUserIsMember(userId);
+                return new ResponseEntity<>(gDTOList, HttpStatus.OK);
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw new InternalServerErrorException("Exception raised trying to retrieve group");
+            }
+        }else{
+            throw new ForbiddenException("user cannot access another users information");
         }
     }
 
@@ -108,17 +121,31 @@ public class Controller {
 
     ////////////////////Controller for groupmembers//////////////
     @PostMapping("/groupmember")
-    public ResponseEntity<GroupMembersDTO> addGroupMemberToGroupById(@RequestBody GroupRequestBody grbm){
-        UserDTO userDTO = userService.getUserById(grbm.getUserId());
-        if (userDTO == null){
-            throw new ResourceNotFoundException("No user with user ID "+grbm.getUserId()+" exists");
+    public ResponseEntity<GroupMembersDTO> addGroupMemberToGroupById(@RequestBody GroupRequestBody grbm, HttpServletRequest request){
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
         } else {
-            try{
-                GroupMembersDTO gmdto = groupService.joinGroup(grbm.getGroupId(), grbm.getUserId(), grbm.isAdmin());
-                return new ResponseEntity<>(gmdto, HttpStatus.CREATED);
-            } catch(Exception e){
-                throw new InternalServerErrorException("User "+grbm.getUserId()+" could not be added to the group");
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+        Long userId = grbm.getUserId();
+
+        if(uid.equals(userId)) {
+            UserDTO userDTO = userService.getUserById(grbm.getUserId());
+            if (userDTO == null){
+                throw new ResourceNotFoundException("No user with user ID "+grbm.getUserId()+" exists");
+            } else {
+                try{
+                    GroupMembersDTO gmdto = groupService.joinGroup(grbm.getGroupId(), grbm.getUserId(), grbm.isAdmin());
+                    return new ResponseEntity<>(gmdto, HttpStatus.CREATED);
+                } catch(Exception e){
+                    throw new InternalServerErrorException("User "+grbm.getUserId()+" could not be added to the group");
+                }
             }
+        }else{
+            throw new ForbiddenException("user cannot access another users information");
         }
     }
 
@@ -135,26 +162,57 @@ public class Controller {
     }
 
     @PutMapping("/groupmember")
-    public ResponseEntity<GroupMembersDTO> updateGroupAdminByUser(@RequestBody GroupRequestBody grbm){
-        GroupMembersDTO gmDTO;
-        try{
-            gmDTO = groupMemberService.updateGroupAdminByUserId(grbm.getUserId(), grbm.getGroupId(), grbm.isAdmin());
-            return new ResponseEntity<>(gmDTO, HttpStatus.OK);
-        } catch(Exception e){
-            //TODO: Dom: Distinguish exception types
-            throw new InternalServerErrorException("Something went wrong, could not update admin");
+    public ResponseEntity<GroupMembersDTO> updateGroupAdminByUser(@RequestBody GroupRequestBody grbm, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId from param:
+        Long userId = grbm.getUserId();
+
+        if (uid.equals(userId)) {
+            GroupMembersDTO gmDTO;
+            try {
+                gmDTO = groupMemberService.updateGroupAdminByUserId(grbm.getUserId(), grbm.getGroupId(), grbm.isAdmin());
+                return new ResponseEntity<>(gmDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                //TODO: Dom: Distinguish exception types
+                throw new InternalServerErrorException("Something went wrong, could not update admin");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     @DeleteMapping(value="/groupmember",params={"uid","gid"})
-    public ResponseEntity<GroupMembersDTO> deleteGroupMemberById(@RequestParam Long uid,@RequestParam Long gid){
-        GroupMembersDTO gmDTO;
-        try{
-            gmDTO = groupMemberService.deleteGroupMemberById(uid, gid);
-            return new ResponseEntity<>(gmDTO, HttpStatus.OK);
-        } catch(Exception e){
-            //TODO: Dom: Distinguish exception types
-            throw new InternalServerErrorException("Something went wrong, could not delete groupmember");
+    public ResponseEntity<GroupMembersDTO> deleteGroupMemberById(@RequestParam Long uid,@RequestParam Long gid, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long userId = getUserIdFromJWT(jwt);
+
+        if (userId.equals(uid)) {
+            GroupMembersDTO gmDTO;
+            try{
+                gmDTO = groupMemberService.deleteGroupMemberById(uid, gid);
+                return new ResponseEntity<>(gmDTO, HttpStatus.OK);
+            } catch(Exception e){
+                //TODO: Dom: Distinguish exception types
+                throw new InternalServerErrorException("Something went wrong, could not delete groupmember");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -176,35 +234,83 @@ public class Controller {
     ////////////////////Controller for friends//////////////
     @PostMapping("/friends")
     @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<FriendDTO> addingFriends(@RequestBody FriendRequestBody ids) {
-        FriendDTO friendDTO;
-        try {
-            friendDTO = friendService.addFriend(ids);
-            return new ResponseEntity<>(friendDTO, HttpStatus.CREATED);
-        } catch (Exception e) {
-            throw new InternalServerErrorException("Exception raised trying to add friend");
+    public ResponseEntity<FriendDTO> addingFriends(@RequestBody FriendRequestBody ids, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId from param:
+        Long userId = ids.getUserId();
+
+        if (uid.equals(userId)) {
+            FriendDTO friendDTO;
+            try {
+                friendDTO = friendService.addFriend(ids);
+                return new ResponseEntity<>(friendDTO, HttpStatus.CREATED);
+            } catch (Exception e) {
+                throw new InternalServerErrorException("Exception raised trying to add friend");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     @DeleteMapping("/friends")
     @ResponseStatus(HttpStatus.OK)
-    public void removeFriend(@RequestBody FriendRequestBody ids) {
-        try {
-            Friend friend = friendService.getFriendById(ids.getUserId(), ids.getFriendId());
-            friendService.deleteFriend(friend);
-        } catch (Exception e) {
-            throw new InternalServerErrorException("Exception raised trying to delete friend");
+    public void removeFriend(@RequestBody FriendRequestBody ids, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId from param:
+        Long userId = ids.getUserId();
+
+        if (uid.equals(userId)) {
+            try {
+                Friend friend = friendService.getFriendById(ids.getUserId(), ids.getFriendId());
+                friendService.deleteFriend(friend);
+            } catch (Exception e) {
+                throw new InternalServerErrorException("Exception raised trying to delete friend");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     @GetMapping(value="/friends", params="userId")
-    public ResponseEntity<List<FriendDTO>> getFriendsOfUser(@RequestParam Long userId) {
-        List<FriendDTO> friendDTOList;
-        try {
-            friendDTOList = friendService.getFriendsByUserId(userId);
-            return new ResponseEntity<>(friendDTOList, HttpStatus.OK);
-        } catch (Exception e) {
-            throw new InternalServerErrorException("Friends for " + userId + " could not be retrieved");
+    public ResponseEntity<List<FriendDTO>> getFriendsOfUser(@RequestParam Long userId, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if (uid.equals(userId)) {
+            List<FriendDTO> friendDTOList;
+            try {
+                friendDTOList = friendService.getFriendsByUserId(userId);
+                return new ResponseEntity<>(friendDTOList, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new InternalServerErrorException("Friends for " + userId + " could not be retrieved");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -221,13 +327,30 @@ public class Controller {
     }
 
     @PostMapping("/posts")
-    public ResponseEntity<PostDTO> addPost(@RequestBody PostRequestBody ids) {
-        try {
-            PostDTO pdto = postService.createPost(ids);
-            return new ResponseEntity<>(pdto, HttpStatus.CREATED);
-        } catch (Exception e) {
-            e.printStackTrace();
-            throw new InternalServerErrorException("Error in creating post: " + e.getMessage());
+    public ResponseEntity<PostDTO> addPost(@RequestBody PostRequestBody ids, HttpServletRequest request) {
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId for param:
+        Long userId = ids.getUserId();
+
+        if (uid.equals(userId)) {
+            try {
+                PostDTO pdto = postService.createPost(ids);
+                return new ResponseEntity<>(pdto, HttpStatus.CREATED);
+            } catch (Exception e) {
+                e.printStackTrace();
+                throw new InternalServerErrorException("Error in creating post: " + e.getMessage());
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -244,13 +367,27 @@ public class Controller {
 
     ////////////////////Controller for users/////////////////////
     @GetMapping(value="/users", params="userId")
-    public ResponseEntity<UserDTO> getUserById(@RequestParam Long userId){
-        UserDTO userDTO;
-        try{
-            userDTO = userService.getUserById(userId);
-            return new ResponseEntity<>(userDTO, HttpStatus.OK);
-        } catch(Exception e){
-            throw new ResourceNotFoundException("No user with user ID "+userId+" exists");
+    public ResponseEntity<UserDTO> getUserById(@RequestParam Long userId, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if (uid.equals(userId)) {
+            UserDTO userDTO;
+            try {
+                userDTO = userService.getUserById(userId);
+                return new ResponseEntity<>(userDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new ResourceNotFoundException("No user with user ID " + userId + " exists");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -265,12 +402,26 @@ public class Controller {
     }
 
     @DeleteMapping(value="/users", params="userId")
-    public ResponseEntity<UserDTO> deleteUserById(@RequestParam Long userId){
-        try{
-            UserDTO userDTO = userService.deleteUserById(userId);
-            return new ResponseEntity<>(userDTO,HttpStatus.OK);
-        } catch(Exception e){
-            throw new ResourceNotFoundException("No user with ID "+userId+" exists");
+    public ResponseEntity<UserDTO> deleteUserById(@RequestParam Long userId, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        if (uid.equals(userId)) {
+            try {
+                UserDTO userDTO = userService.deleteUserById(userId);
+                return new ResponseEntity<>(userDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new ResourceNotFoundException("No user with ID " + userId + " exists");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
@@ -288,17 +439,33 @@ public class Controller {
     }
 
     @PutMapping(value="/users")
-    public ResponseEntity<UserDTO> updateUser(@RequestBody UpdateDispnameRequestBody udrb){
-        try{
-            UserDTO userDTO = userService.updateUser(udrb);
-            return new ResponseEntity<>(userDTO, HttpStatus.OK);
-        } catch(Exception e){
-            throw new ResourceNotFoundException("No such user exists");
+    public ResponseEntity<UserDTO> updateUser(@RequestBody UpdateDispnameRequestBody udrb, HttpServletRequest request){
+        //Get userId from token:
+        String token = request.getHeader("Authorization");
+        String jwt = null;
+        if(token.startsWith("Bearer ")){
+            jwt = token.substring(7);
+        } else {
+            throw new ResourceNotFoundException();
+        }
+        Long uid = getUserIdFromJWT(jwt);
+
+        //Get userId for param:
+        Long userId = udrb.getUserid();
+
+        if (uid.equals(userId)) {
+            try {
+                UserDTO userDTO = userService.updateUser(udrb);
+                return new ResponseEntity<>(userDTO, HttpStatus.OK);
+            } catch (Exception e) {
+                throw new ResourceNotFoundException("No such user exists");
+            }
+        } else {
+            throw new ForbiddenException("User cannot access another users information");
         }
     }
 
     ////////////////////Register, logout/////////////////////
-
     @PostMapping(value="/register")
     public ResponseEntity<CreateUserStatus> registerNewUser(@RequestBody CreateUserRequestBody params){
         try{
